requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
api==0.0.7
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.6
distlib==0.3.8
distro==1.9.0
faiss-cpu==1.8.0
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.60
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
#networkx==3.3
nose==1.3.7
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
pyarrow==16.1.0
PyAudio==0.2.14
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.1
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
streamlit==1.34.0
sympy==1.12
tenacity==8.3.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
yarl==1.9.4
streamlit
python-dotenv
openai
pynput
time
streamlit-audiorecorder
transformers